【modify】 首推ID 修改为整形

+++ zyx-mapper/src/main/java/com/zyx/mapper/pg/CircleMapper.java
@@ -70,6 +70,6 @@ public interface CircleMapper extends BaseMapper<Circle> {
      * @param ids
      * @return
      */
-    List<Circle> queryCircleDeva(@Param("ids") List<Long> ids);
+    List<Circle> queryCircleDeva(@Param("ids") List<Integer> ids);
 
 }
